+++ bigquery/tests/system.py
@@ -48,6 +48,7 @@ from google.api_core.exceptions import PreconditionFailed
 from google.api_core.exceptions import BadRequest
 from google.api_core.exceptions import Conflict
 from google.api_core.exceptions import Forbidden
+from google.api_core.exceptions import GoogleAPICallError
 from google.api_core.exceptions import NotFound
 from google.api_core.exceptions import InternalServerError
 from google.api_core.exceptions import ServiceUnavailable
@@ -765,17 +766,17 @@ class TestBigQuery(unittest.TestCase):
         self.assertTrue(eu_query.done)
 
         # Cannot query from US.
-        with self.assertRaises(BadRequest):
+        with self.assertRaises(GoogleAPICallError):
             list(client.query(query_string, location="US", job_config=query_config))
 
         # Cannot copy from US.
-        with self.assertRaises(BadRequest):
+        with self.assertRaises(GoogleAPICallError):
             client.copy_table(
                 table_ref, dataset.table("letters2_us"), location="US"
             ).result()
 
         # Cannot extract from US.
-        with self.assertRaises(BadRequest):
+        with self.assertRaises(GoogleAPICallError):
             client.extract_table(
                 table_ref, "gs://{}/letters-us.csv".format(bucket_name), location="US"
             ).result()
@@ -1733,13 +1734,22 @@ class TestBigQuery(unittest.TestCase):
                     ),
                 ],
             ),
+            SF("bigfloat_col", "FLOAT", mode="NULLABLE"),
+            SF("smallfloat_col", "FLOAT", mode="NULLABLE"),
         ]
         record = {
             "nested_string": "another string value",
             "nested_repeated": [0, 1, 2],
             "nested_record": {"nested_nested_string": "some deep insight"},
         }
-        to_insert = [{"string_col": "Some value", "record_col": record}]
+        to_insert = [
+            {
+                "string_col": "Some value",
+                "record_col": record,
+                "bigfloat_col": 3.14,
+                "smallfloat_col": 2.72,
+            }
+        ]
         rows = [json.dumps(row) for row in to_insert]
         body = six.BytesIO("{}\n".format("\n".join(rows)).encode("ascii"))
         table_id = "test_table"
@@ -1753,11 +1763,13 @@ class TestBigQuery(unittest.TestCase):
         # Load a table using a local JSON file from memory.
         Config.CLIENT.load_table_from_file(body, table, job_config=job_config).result()
 
-        df = Config.CLIENT.list_rows(table, selected_fields=schema).to_dataframe()
+        df = Config.CLIENT.list_rows(table, selected_fields=schema).to_dataframe(
+            dtypes={"smallfloat_col": "float16"}
+        )
 
         self.assertIsInstance(df, pandas.DataFrame)
         self.assertEqual(len(df), 1)  # verify the number of rows
-        exp_columns = ["string_col", "record_col"]
+        exp_columns = ["string_col", "record_col", "bigfloat_col", "smallfloat_col"]
         self.assertEqual(list(df), exp_columns)  # verify the column names
         row = df.iloc[0]
         # verify the row content
@@ -1769,6 +1781,9 @@ class TestBigQuery(unittest.TestCase):
             row["record_col"]["nested_record"]["nested_nested_string"],
             "some deep insight",
         )
+        # verify dtypes
+        self.assertEqual(df.dtypes["bigfloat_col"].name, "float64")
+        self.assertEqual(df.dtypes["smallfloat_col"].name, "float16")
 
     def test_list_rows_empty_table(self):
         from google.cloud.bigquery.table import RowIterator