+++ redis/synth.py
@@ -29,7 +29,10 @@ excludes = ["README.rst", "setup.py", "nox*.py", "docs/conf.py", "docs/index.rst
 # ----------------------------------------------------------------------------
 for version in ["v1beta1", "v1"]:
     library = gapic.py_library(
-        "redis", version, config_path=f"artman_redis_{version}.yaml"
+        "redis",
+        version,
+        config_path=f"artman_redis_{version}.yaml",
+        include_protos=True,
     )
 
     s.copy(library, excludes=excludes)